requirements.txt
langchain
langchain-core
langchain-community
langchain-mistralai
langchain-huggingface
faiss-cpu
sentence-transformers
transformers
torch
python-dotenv
streamlit
uvicorn
fastapi
python-dotenv
